+++ src/Concordium/Client/Runner.hs
@@ -198,6 +198,7 @@ process Options{optsCmd = command, optsBackend = backend, optsConfigDir = cfgDir
         BakerCmd c -> processBakerCmd c cfgDir verbose backend
         DelegatorCmd c -> processDelegatorCmd c cfgDir verbose backend
         IdentityCmd c -> processIdentityCmd c backend
+        PoolCmd c -> processPoolCmd c backend
 
 -- | Process a 'config ...' command.
 processConfigCmd :: ConfigCmd -> Maybe FilePath -> Verbose -> IO ()
@@ -4113,6 +4114,23 @@ processIdentityShowCmd action backend =
                 getAnonymityRevokers bhInput
                     >>= getResponseValueOrDie
                     >>= runPrinter . printAnonymityRevokers . toList
+
+-- | Process a 'pool ...' command.
+processPoolCmd :: PoolCmd -> Backend -> IO ()
+processPoolCmd action backend =
+    case action of
+        PoolStatus pool block -> do
+            withClient backend $ do
+                b <- readBlockHashOrDefault Best block
+                case pool of
+                    Nothing -> do
+                        getPassiveDelegationInfo b
+                            >>= getResponseValueOrDie
+                            >>= runPrinter . printPassiveDelegationInfo
+                    Just p -> do
+                        getPoolInfo b p
+                            >>= getResponseValueOrDie
+                            >>= runPrinter . printPoolInfo
 
 -- | Process a "legacy" command.
 processLegacyCmd :: LegacyCmd -> Backend -> IO ()