+++ ChangeLog.md
@@ -4,6 +4,11 @@
 
 - The `transaction status` command output includes the `parameter` for
   `ContractInitialized` events.
+- From protocol version 8, raw command `GetPoolStatus` indicates if a validator is suspended and,
+  if it is in the current committee, if it is primed for suspension and the current count of
+  missed rounds.
+- From protocol version 8 `account show` also shows if a validator is suspended.
+- Add command `pool status` for getting information about a pool.
 
 ## 8.0.0
 
@@ -20,9 +25,6 @@ Note: due to API changes, this release may not work correctly with node versions
   accounts with scheduled releases, cooldowns, pre-cooldowns and pre-pre-cooldowns.
 - Raw commands `GetBlockTransactionEvents` and `GetTransactionStatus` include the `parameter`
   for `ContractInitialized` events.
-- From protocol version 8, raw command `GetPoolStatus` indicates if a validator is suspended and,
-  if it is in the current committee, if it is primed for suspension and the current count of
-  missed rounds.
 
 ## 7.0.1